+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # ocapld ChangeLog
 
+### Changed
+- Use webpack 'externals' for jsonld and jsonld-signatures.
+
 ## 0.1.0 - 2019-01-02
 - Initial release.
 - See git history for changes.